'Refactored by Sourcery'

+++ lynda/modules/cleaner.py
@@ -10,11 +10,7 @@ from lynda.modules.disable import DisableAbleCommandHandler
 from lynda.modules.helper_funcs.chat_status import user_admin, bot_can_delete, dev_plus, connection_status
 from lynda.modules.sql import cleaner_sql as sql
 
-if ALLOW_EXCL:
-    CMD_STARTERS = ('/', '!')
-else:
-    CMD_STARTERS = ('/')
-
+CMD_STARTERS = ('/', '!') if ALLOW_EXCL else '/'
 BLUE_TEXT_CLEAN_GROUP = 15
 CommandHandlerList = (CommandHandler, CustomCommandHandler, DisableAbleCommandHandler)
 command_list = ["cleanbluetext", "ignorecleanbluetext", "unignorecleanbluetext", "listcleanbluetext", "ignoreglobalcleanbluetext", "unignoreglobalcleanbluetext"
@@ -32,21 +28,19 @@ def clean_blue_text_must_click(bot: Bot, update: Update):
     chat = update.effective_chat
     message = update.effective_message
 
-    if chat.get_member(bot.id).can_delete_messages:
-        if sql.is_enabled(chat.id):
-            fst_word = message.text.strip().split(None, 1)[0]
+    if chat.get_member(bot.id).can_delete_messages and sql.is_enabled(chat.id):
+        fst_word = message.text.strip().split(None, 1)[0]
 
-            if len(fst_word) > 1 and any(fst_word.startswith(start) for start in CMD_STARTERS):
+        if len(fst_word) > 1 and any(fst_word.startswith(start) for start in CMD_STARTERS):
 
-                command = fst_word[1:].split('@')
-                chat = update.effective_chat
+            command = fst_word[1:].split('@')
+            chat = update.effective_chat
 
-                ignored = sql.is_command_ignored(chat.id, command[0])
-                if ignored:
-                    return
+            if ignored := sql.is_command_ignored(chat.id, command[0]):
+                return
 
-                if command[0] not in command_list:
-                    message.delete()
+            if command[0] not in command_list:
+                message.delete()
 
 
 @run_async
@@ -58,16 +52,16 @@ def set_blue_text_must_click(bot: Bot, update: Update, args: List[str]):
     chat = update.effective_chat
     message = update.effective_message
 
-    if len(args) >= 1:
+    if args:
         val = args[0].lower()
-        if val == "off" or val == "no":
+        if val in ["off", "no"]:
             sql.set_cleanbt(chat.id, False)
-            reply = "Bluetext cleaning has been disabled for <b>{}</b>".format(html.escape(chat.title))
+            reply = f"Bluetext cleaning has been disabled for <b>{html.escape(chat.title)}</b>"
             message.reply_text(reply, parse_mode=ParseMode.HTML)
 
-        elif val == "yes" or val == "on":
+        elif val in ["yes", "on"]:
             sql.set_cleanbt(chat.id, True)
-            reply = "Bluetext cleaning has been enabled for <b>{}</b>".format(html.escape(chat.title))
+            reply = f"Bluetext cleaning has been enabled for <b>{html.escape(chat.title)}</b>"
             message.reply_text(reply, parse_mode=ParseMode.HTML)
 
         else:
@@ -75,11 +69,8 @@ def set_blue_text_must_click(bot: Bot, update: Update, args: List[str]):
             message.reply_text(reply)
     else:
         clean_status = sql.is_enabled(chat.id)
-        if clean_status:
-            clean_status = "Enabled"
-        else:
-            clean_status = "Disabled"
-        reply = "Bluetext cleaning for <b>{}</b> : <b>{}</b>".format(chat.title, clean_status)
+        clean_status = "Enabled" if clean_status else "Disabled"
+        reply = f"Bluetext cleaning for <b>{chat.title}</b> : <b>{clean_status}</b>"
         message.reply_text(reply, parse_mode=ParseMode.HTML)
 
 
@@ -90,15 +81,14 @@ def add_bluetext_ignore(bot: Bot, update: Update, args: List[str]):
     message = update.effective_message
     chat = update.effective_chat
 
-    if len(args) >= 1:
+    if args:
         val = args[0].lower()
-        added = sql.chat_ignore_command(chat.id, val)
-        if added:
-            reply = "<b>{}</b> has been added to bluetext cleaner ignore list.".format(args[0])
+        if added := sql.chat_ignore_command(chat.id, val):
+            reply = f"<b>{args[0]}</b> has been added to bluetext cleaner ignore list."
         else:
             reply = "Command is already ignored."
         message.reply_text(reply, parse_mode=ParseMode.HTML)
-        
+
     else:
         reply = "No command supplied to be ignored."
         message.reply_text(reply)
@@ -111,15 +101,14 @@ def remove_bluetext_ignore(bot: Bot, update: Update, args: List[str]):
     message = update.effective_message
     chat = update.effective_chat
 
-    if len(args) >= 1:
+    if args:
         val = args[0].lower()
-        removed = sql.chat_unignore_command(chat.id, val)
-        if removed:
-            reply = "<b>{}</b> has been removed from bluetext cleaner ignore list.".format(args[0])
+        if removed := sql.chat_unignore_command(chat.id, val):
+            reply = f"<b>{args[0]}</b> has been removed from bluetext cleaner ignore list."
         else:
             reply = "Command isn't ignored currently."
         message.reply_text(reply, parse_mode=ParseMode.HTML)
-        
+
     else:
         reply = "No command supplied to be unignored."
         message.reply_text(reply)
@@ -131,15 +120,14 @@ def add_bluetext_ignore_global(bot: Bot, update: Update, args: List[str]):
 
     message = update.effective_message
 
-    if len(args) >= 1:
+    if args:
         val = args[0].lower()
-        added = sql.global_ignore_command(val)
-        if added:
-            reply = "<b>{}</b> has been added to global bluetext cleaner ignore list.".format(args[0])
+        if added := sql.global_ignore_command(val):
+            reply = f"<b>{args[0]}</b> has been added to global bluetext cleaner ignore list."
         else:
             reply = "Command is already ignored."
         message.reply_text(reply, parse_mode=ParseMode.HTML)
-        
+
     else:
         reply = "No command supplied to be ignored."
         message.reply_text(reply)
@@ -151,15 +139,14 @@ def remove_bluetext_ignore_global(bot: Bot, update: Update, args: List[str]):
 
     message = update.effective_message
 
-    if len(args) >= 1:
+    if args:
         val = args[0].lower()
-        removed = sql.global_unignore_command(val)
-        if removed:
-            reply = "<b>{}</b> has been removed from global bluetext cleaner ignore list.".format(args[0])
+        if removed := sql.global_unignore_command(val):
+            reply = f"<b>{args[0]}</b> has been removed from global bluetext cleaner ignore list."
         else:
             reply = "Command isn't ignored currently."
         message.reply_text(reply, parse_mode=ParseMode.HTML)
-        
+
     else:
         reply = "No command supplied to be unignored."
         message.reply_text(reply)
